Notes from Buddy1
The code itself is fine and the application (both tabs) open and run properly. I don't see any code for the layout portion of the assignment and I am unable to get my tablet generator to run so I can't test the proportions. Just a few thoughts about the format. You could make the font a little larger on the initial page and perhaps bold your name. You could also offset the image and/or the text with a box that has a different colored background (probably white).